Prepare the refresher for HTTPS
